+++ docs/userguide/project/project.py
@@ -142,7 +142,7 @@ proj2
 _ = proj2.NMR_1D.plot()
 
 # %%
-proj2.NMR_2D  #  TODO: something wrong there!
+proj2.NMR_2D
 
 # %%
 proj.NMR_2D.plot()