+++ CHANGELOG.md
@@ -4,13 +4,15 @@
 
 **NEW FEATURES**
 
-* a default coordiate can now be selected for multiple coordinates dimensions.
+* A default coordinate can now be selected for multiple coordinates dimensions.
 
 **BUGS FIXED**
 
-* alignment along several dimensions (issue #248)
+* Alignment along several dimensions (issue #248)
 
 * to() and ito() methods have been fixed to work correctly (issue #255) 
+
+* Baseline correction works on all dimensions 
 
 ## Version 0.2.13
 
@@ -25,6 +27,7 @@
 * LinearCoord operations now working.
 
 * Baseline default now "sequential" as expected. 
+  
   **WARNING**: It was wrongly set to "mutivariate" in previous releases, so you should 
   expect some difference with processings you may have done before.